requirement.txt
ddddocr==1.4.1
jsonpath==0.82
numpy==1.21.5
onnxruntime==1.10.0
opencv-python==4.5.5.62
Pillow==9.0.0
PyExecJS==1.5.1
Scrapy==2.5.1

